Update
Smart contract playbook was having issue with the blockchain because the previous async task is still running in the back ground so gave the previous task and extra 10 second wait time.

Updated the ansible-dappley.go program so instead of manually typing out the files that needs to be updated it automatically detects and updates.